Backend: added and renamed some directories
Added domain and dto directories among others, to follow conventional
structure
