+++ backend/src/main/java/dat251_gruppe2/hotelifinder/domain/exampleModel.java
@@ -1,0 +1,33 @@
+package dat251_gruppe2.hotelifinder.domain;
+
+public class exampleModel {
+}
+
+/*
+ * Example of User domain/class:
+ * 
+ * @Entity
+ * 
+ * @Table(name="users") // User is a reserved table name
+ * public class User {
+ * 
+ * @Id
+ * 
+ * @GeneratedValue(strategy = GenerationType.IDENTITY)
+ * private Long id;
+ * private String username;
+ * 
+ * 
+ * public User() {
+ * this.username = "";
+ * }
+ * 
+ * public String getUsername() {
+ * return username;
+ * }
+ * 
+ * public void setUsername(String username) {
+ * this.username = username;
+ * }
+ * }
+ */
