--- backend/src/main/java/dat251_gruppe2/hotelifinder/model/exampleModel.java
@@ -1,32 +1,0 @@
-package dat251_gruppe2.hotelifinder.model;
-
-public class exampleModel {
-}
-
-
-/*
-Example of User domain/class:
-
-@Entity
-@Table(name="users") // User is a reserved table name
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String username;
-
-
-    public User() {
-        this.username = "";
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-}
- */
