+++ backend/src/main/java/dat251_gruppe2/hotelifinder/repositories/exampleRepository.java
@@ -1,0 +1,15 @@
+package dat251_gruppe2.hotelifinder.repositories;
+
+public interface exampleRepository {
+}
+
+/*
+ * Example:
+ * 
+ * public interface PollRepository extends JpaRepository<Poll, Long> {
+ * 
+ * Poll save(Poll poll);
+ * 
+ * void deleteById(Long id);
+ * }
+ */
