--- backend/src/main/java/dat251_gruppe2/hotelifinder/repository/exampleRepository.java
@@ -1,16 +1,0 @@
-package dat251_gruppe2.hotelifinder.repository;
-
-
-public interface exampleRepository {
-}
-
-/*
-Example:
-
-public interface PollRepository extends JpaRepository<Poll, Long> {
-
-    Poll save(Poll poll);
-
-    void deleteById(Long id);
-}
-*/
