+++ backend/src/main/java/dat251_gruppe2/hotelifinder/services/exampleService.java
@@ -1,8 +1,6 @@
-package dat251_gruppe2.hotelifinder.service;
+package dat251_gruppe2.hotelifinder.services;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public class exampleService {
 }
-
-
